Only build when new changes are commited
With this implemented, build.sh will only build if the count of commits
from all projects aren't equial, it will create a file to save the
amount of commits by device and branch.
There will be a flag to ignore this check, if $IGNORE_COMMIT_COUNT=true
this check is avoided.
File created to save the commit count by device/branch
.${DEVICE}_${BRANCH}_COMMIT_COUNT

Change-Id: I34ca462d52d1b2ae10978221346cae86d46e34be
Signed-off-by: André Pinela <sheffzor@gmail.com>